fastapi-the-complete-course/project3/TodoApp/pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
]
[project.scripts]
start = "uvicorn main:app --reload"
